+++ docs/source/whatsnew.rst
@@ -4,6 +4,7 @@ What's New
 ==========
 v3.1.0
 ------
+- XSI now offers a :py:meth:`directory()` method to search Enterprise, Group and Personal directories.
 - PERFORMANCE: **wxcadm** previously retrieved too much data when Orgs and Locations were initialized. The API calls have been moved so they are only performed when the relevant data is needed.
 - PERFORMANCE: :py:meth:`Location.numbers` and :py:meth:`Location.asailable_numbers` now don't pull all numbers for the Org, only the Location
 - **wxcadm** now supports Wholesale partners with the :py:class:`WholesaleCustomer` and :py:class:`WholesaleSubscriber` classes